First stab at getting this to work :)
Use Parse::CPAN::Packages to get our data, then present it via
MojoX::Renderer::YAML.  Have a test script done, as well.

02packages.details.txt.gz is nearby, and probably needs a model class so
it can be abstracted away.  For now, however, we can use plackup to
check for updates to this file so the app can be reloaded when new
modules come in to the database.

Signed-off-by: Zak B. Elep <zakame@zakame.net>